Fix mach.subcode always 0 due to 32-bit/64-bit struct mismatch (#810) (#811)
The Mach exception handler registered with MACH_EXCEPTION_CODES (64-bit
codes) but decoded the request using __Request__exception_raise_t which
has 32-bit integer_t code[2]. This caused code[1] to read the upper half
of code[0] instead of the actual subcode, producing 0 for every crash.

Replace ExceptionRequest with a correct 64-bit struct matching the
mach_exc MIG layout, centralize decode in machCodeFromRequest/
machSubcodeFromRequest helpers, and fix the report writer casts that
also truncated to 32 bits.

+++ Sources/KSCrashRecording/KSCrashReportC.c
@@ -1365,11 +1365,11 @@ static void writeError(const KSCrashReportWriter *const writer, const char *cons
             if (machExceptionName != NULL) {
                 writer->addStringElement(writer, KSCrashField_ExceptionName, machExceptionName);
             }
-            writer->addUIntegerElement(writer, KSCrashField_Code, (unsigned)crash->mach.code);
+            writer->addUIntegerElement(writer, KSCrashField_Code, (uint64_t)crash->mach.code);
             if (machCodeName != NULL) {
                 writer->addStringElement(writer, KSCrashField_CodeName, machCodeName);
             }
-            writer->addUIntegerElement(writer, KSCrashField_Subcode, (size_t)crash->mach.subcode);
+            writer->addUIntegerElement(writer, KSCrashField_Subcode, (uint64_t)crash->mach.subcode);
         }
         writer->endContainer(writer);
 #endif
@@ -1863,4 +1863,3 @@ void kscrashreport_setIsWritingReportCallback(const KSCrashIsWritingReportCallba
 }
 
 void kscrashreport_setCompactBinaryImages(bool enabled) { g_compactBinaryImages = enabled; }
-
